backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
supabase
python-jose[cryptography]
passlib[bcrypt]
httpx<0.28
aiosmtplib
gunicorn
pydantic-settings
pydantic[email]
python-multipart
slowapi
